Display total earnings in wallet sidebar
