+++ src/Pages/Authenticated/Wallet/TotalMystCard.scss
@@ -1,0 +1,34 @@
+@import 'src/assets/styles/base/variables';
+
+.myst-and-data {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+
+  &__footer {
+    &__sum-block {
+      width: 100%;
+      text-align: center;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      padding: 15px 0;
+
+      .name {
+        font-weight: $medium-weight;
+        font-size: 16px;
+        line-height: 24px;
+        color: $main-text-color;
+      }
+
+      .explanation {
+        font-style: normal;
+        font-weight: $normal-weight;
+        font-size: 14px;
+        line-height: 21px;
+        color: $names-text-color;
+      }
+    }
+  }
+}
